+++ connector_amazon/models/amazon_sale_importer.py
@@ -1,13 +1,14 @@
 # -*- coding: utf-8 -*-
 # Copyright 2017 Akretion (http://www.akretion.com).
 # @author Sébastien BEAU <sebastien.beau@akretion.com>
+# @author David BEAL <david.beal@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 import StringIO
 import logging
 
-from openerp import _, models
-from openerp.exceptions import Warning as UserError
+from openerp import models
+# from openerp.exceptions import Warning as UserError
 
 _logger = logging.getLogger(__name__)
 
@@ -35,8 +36,8 @@ class AmazonSaleImporter(models.AbstractModel):
         sales = self._extract_infos(reader)
         file.close()
         backend = meta_attachment.amazon_backend_id
-        for order_item, sale in sales.items():
-            self._create_sale(sale, order_item, meta_attachment, backend)
+        for item in sales:
+            backend._create_sale(sales[item], meta_attachment)
 
     def _get_header_fieldnames(self):
         return [
@@ -97,115 +98,3 @@ class AmazonSaleImporter(models.AbstractModel):
             'shipping': float(line['shipping-price']) + \
             float(line['shipping-tax']),
         }
-
-    def _create_sale(self, sale, order_item, meta_attachment, backend):
-        """ We process sale order of the file
-        """
-        partner = self._get_customer(sale['partner'])
-        part_ship = self._get_delivery_address(
-            sale['part_ship'], sale['sale']['origin'], partner)
-        vals = {
-            'name': (backend.sale_prefix or '') + sale['sale']['origin'],
-            'partner_id': partner.id,
-            'partner_shipping_id': part_ship.id,
-            'pricelist_id': backend.pricelist_id.id,
-            'external_origin': 'ir.attachment.metadata,%s'
-            % meta_attachment.id,
-            'origin': sale['sale']['origin'],
-        }
-        ship_price = self._prepare_products(sale['lines'], backend)
-        vals['order_line'] = [
-            (0, 0, {key: val for key, val in line.items()
-                    if key in self.env['sale.order.line']._fields.keys()})
-            for line in sale['lines']
-        ]
-        if ship_price:
-            ship_vals = {
-                'product_uom_qty': 1,
-                'price_unit': ship_price,
-                'product_id': backend.shipping_product.id,
-            }
-            vals['order_line'].append((0, 0, ship_vals), )
-        self.env['sale.order'].create(vals)
-
-    def _prepare_products(self, lines, backend):
-        """ - check if product exist in amazon backend
-            - gather shipping price
-            return shipping_price
-        """
-        line_count, shipping_price = 0, 0
-        products_in_exception = []
-        for line in lines:
-            shipping_line = float(line.get('shipping'))
-            if shipping_line:
-                shipping_price += shipping_line
-            binding = self.env['amazon.product'].search(
-                [('external_id', '=', line['sku']),
-                 ('backend_id', '=', backend.id)])
-            if binding:
-                lines[line_count]['product_id'] = binding[0].record_id.id
-            else:
-                products_in_exception.append(line['sku'])
-            line_count += 1
-        if products_in_exception:
-            raise UserError(
-                _("No matching product with these sku '%s' in Amazon binding"
-                  % ', '.join(products_in_exception)))
-        return shipping_price
-
-    def _get_customer(self, customer_data):
-        partner_m = self.env['res.partner']
-        partner = partner_m.search(
-            [('email', '=', customer_data['email'])])
-        if not partner:
-            partner = partner_m.create(customer_data)
-        return partner[0]
-
-    def _get_delivery_address(self, part_ship, origin, partner):
-        partner_m = self.env['res.partner']
-        self._prepare_address(part_ship, origin)
-        domain = [
-            '|',
-            ('active', '=', True),
-            ('active', '=', False)]
-        domain.extend([
-            (fieldname, '=', val)
-            for fieldname, val in part_ship.items()
-            if fieldname in partner_m._fields])
-        # we search identical partner active or not
-        address = partner_m.search(domain)
-        if not address:
-            part_ship['parent_id'] = partner.id
-            vals = {k: v for k, v in part_ship.items()
-                    if k in partner_m._fields}
-            address = partner_m.create(vals)
-        return address[0]
-
-    def _prepare_address(self, part_ship, origin):
-        partner_m = self.env['res.partner']
-        part_ship['country_id'], part_ship['state_id'] = \
-            self._get_state_country(part_ship, origin)
-        if part_ship.get('street3') and 'street3' not in partner_m._fields:
-            # if street3 doesn't exist in odoo: according to modules
-            part_ship['street2'] = '%s %s' % (
-                part_ship['street2'], part_ship['street3'])
-
-    def _get_state_country(self, part_ship, origin):
-        """ country is mandatory, not state
-        """
-        country_code = part_ship.get('country')
-        state_name = part_ship.get('state')
-        country = self.env['res.country'].search([('code', '=', country_code)])
-        if not country or not country_code:
-            raise UserError(
-                _("Unknow country code %s in sale %s ") % (
-                    country_code, origin))
-        state = False
-        if state_name:
-            state = self.env['res.country.state'].search(
-                [('name', '=', state_name)])
-            if not state:
-                raise UserError(
-                    _("Unknown state code %s in sale %s ") % (
-                        state_name, origin))
-        return(country.id, getattr(state, 'id', state))
